requirements.txt
Django~=5.0.3
python-statemachine
